added the function to replace the tenant id automatically in request URL
replace the second string in the url with the tenant id. For example, if
there is an URL looks like this:
    http://hostname:8080/part1/part2/part3?p1=pv1&p2=pv2

with this feature, the string 'part2' in the url will be replaced with
the tenant id automatically. To turn off this feature, set the config
parameter sub_tenant_id = False.